wire up controller
