+++ admin/controllers/access-arrangements.js
@@ -209,35 +209,4 @@ controller.getDeleteAccessArrangements = async (req, res, next) => {
   return res.redirect(`/access-arrangements/overview?hl=${pupil.urlSlug}`)
 }
 
-controller.getAddRetroInputAssistant = async (req, res, next) => {
-  res.locals.pageTitle = 'Record input assistant used for official check'
-  req.breadcrumbs('Select pupils and access arrangements', 'select-access-arrangements')
-  req.breadcrumbs('Record input assistant')
-
-  try {
-    const checkWindowData = await checkWindowV2Service.getActiveCheckWindow()
-    await businessAvailabilityService.determineAccessArrangementsEligibility(checkWindowData)
-  } catch (error) {
-    return next(error)
-  }
-
-  let pupils
-  try {
-    pupils = await pupilAccessArrangementsService.getEligiblePupilsWithFullNames(req.user.schoolId)
-  } catch (error) {
-    return next(error)
-  }
-  return res.render('access-arrangements/retro-add-input-assistant', {
-    breadcrumbs: req.breadcrumbs(),
-    pupils,
-    error: new ValidationError(),
-    formData: {}
-  })
-}
-
-controller.postSubmitRetroInputAssistant = async (req, res, next) => {
-  console.dir(req.body)
-  return res.send(501)
-}
-
 module.exports = controller
