+++ admin/services/data-access/retro-input-assistant.data.service.js
@@ -16,7 +16,6 @@ const service = {
    */
 
   /**
-   *
    * @param {retroInputAssistantData} data
    * @returns {Promise<any>}
    */
